feat: CREATE, UPDATE, DELETEに認証処理を追加

+++ services/item_service.go
@@ -1,19 +1,18 @@
 package services
 
 import (
+	"errors"
 	"learning-fleamarket-api/dto"
 	"learning-fleamarket-api/models"
 	"learning-fleamarket-api/repositories"
 )
 
-// 商品IDの検索、更新、削除に認証を追加
-
 type IItemService interface {
 	FindAll() (*[]models.Item, error)
 	FindByID(itemId uint) (*models.Item, error)
 	Create(createItemInput dto.CreateItemInput, userId uint) (*models.Item, error)
-	Update(itemId uint, updateItemInput dto.UpdateItemInput) (*models.Item, error)
-	Delete(itemId uint) error
+	Update(itemId uint, updateItemInput dto.UpdateItemInput, userId uint) (*models.Item, error)
+	Delete(itemId uint, userId uint) error
 }
 
 type ItemService struct {
@@ -47,10 +46,13 @@ func (s *ItemService) Create(createItemInput dto.CreateItemInput, userId uint) (
 	return s.repository.Create(newItem)
 }
 
-func (s *ItemService) Update(itemId uint, updateItemInput dto.UpdateItemInput) (*models.Item, error) {
+func (s *ItemService) Update(itemId uint, updateItemInput dto.UpdateItemInput, userId uint) (*models.Item, error) {
 	targetItem, err := s.FindByID(itemId)
 	if err != nil {
 		return nil, err
+	}
+	if targetItem.UserID != userId {
+		return nil, errors.New("Unauthorized error")
 	}
 
 	if updateItemInput.Name != nil {
@@ -68,6 +70,13 @@ func (s *ItemService) Update(itemId uint, updateItemInput dto.UpdateItemInput) (
 	return s.repository.Update(*targetItem)
 }
 
-func (s *ItemService) Delete(itemId uint) error {
+func (s *ItemService) Delete(itemId uint, userId uint) error {
+	targetItem, err := s.FindByID(itemId)
+	if err != nil {
+		return err
+	}
+	if targetItem.UserID != userId {
+		return errors.New("Unauthorized error")
+	}
 	return s.repository.Delete(itemId)
 }
